requirements.txt
fabric
jinja2
pyyaml
ansible-core
